pyproject.toml
[project]
name = "aiterm"
version = "0.1.0"
description = "AI-powered terminal command assistant"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.2.0",
    "openai>=1.78.1",
    "anthropic>=0.18.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "rich>=14.0.0",
    "aiohttp>=3.11.18",
]

[project.scripts]
at = "aiterm.main:main"

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages.find.where = ["src"]
